code/backend/requirements.txt
# server requirement
flask
Werkzeug

# video transcriptions  requirement
moviepy
SpeechRecognition
python-docx
openai


# api requirement
fastapi
uvicorn
pydantic

# database requirement
SQLAlchemy
mysql-connector-python


python-dotenv



# Werkzeug==2.0.3

# absl-py
# astunparse
# cachetools
# certifi
# charset-normalizer
# click
# colorama
# dnspython
# filelock
# Flask /
# flatbuffers
# gast
# google-auth
# google-auth-oauthlib
# google-pasta
# grpcio
# gunicorn
# h5py
# huggingface-hub
# idna
# importlib-metadata
# itsdangerous
# Jinja2
# joblib
# keras
# Keras-Preprocessing
# libclang
# Markdown
# MarkupSafe
# networkx
# nltk
# # numpy
# oauthlib
# opt-einsum
# packaging
# protobuf
# pyasn1
# pyasn1-modules
# pydub
# pymongo
# pyparsing
# PyYAML
# regex
# requests
# requests-oauthlib
# rsa
# sacremoses
# scipy
# sentencepiece
# six
# SpeechRecognition
# tensorboard
# tensorboard-data-server
# tensorboard-plugin-wit
# tensorflow
# tensorflow-io-gcs-filesystem
# termcolor
# tf-estimator-nightly
# tokenizers
# torch
# tqdm
# transformers
# typing-extensions
# urllib3
# Werkzeug
# wrapt
# zipp
